refactor: Move theme options to one file
