+++ src/config/theme.ts
@@ -1,5 +1,21 @@
 import { createTheme } from '@mui/material/styles';
 
+declare module '@mui/material/styles' {
+    interface BreakpointOverrides {
+        xs: false;
+        sm: false;
+        md: false;
+        lg: false;
+        xl: false;
+        phone: true;
+        tablet: true;
+        tabletLandscape: true;
+        laptop: true;
+        desktop: true;
+        desktopWide: true;
+    }
+}
+
 export const theme = createTheme({
     palette: {
         primary: {
@@ -9,6 +25,16 @@ export const theme = createTheme({
         },
         background: {
             default: '#1A1B22'
+        }
+    },
+    breakpoints: {
+        values: {
+            phone: 483,
+            tablet: 640,
+            tabletLandscape: 900,
+            laptop: 1024,
+            desktop: 1501,
+            desktopWide: 1800
         }
     }
 });
