+++ src/hooks/useOurMedia/index.tsx
@@ -1,2 +1,1 @@
-export * from './theme';
 export * from './useOurMedia';
